tokenError

+++ Mern/frontend/src/components/WorkoutDetails.js
@@ -1,15 +1,21 @@
 import React from "react";
-
 import { useWorkoutContext } from "../hooks/useWorkoutContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const WorkoutDetails = ({ workout }) => {
   const { title, reps, load, createdAt } = workout;
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
 
   const deleteHandler = async () => {
+    if (!user) {
+      return;
+    }
+
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: { Authorization: `Bearer ${user.json}` },
     });
     const json = await response.json();
